--- app/src/main/java/woowacourse/paint/PaintActionHandler.kt
@@ -1,7 +1,0 @@
-package woowacourse.paint
-
-import androidx.annotation.ColorRes
-
-interface PaintActionHandler {
-    fun changeColorRes(@ColorRes colorRes: Int)
-}
