--- app/src/main/java/woowacourse/paint/PaintColor.kt
@@ -1,3 +1,0 @@
-package woowacourse.paint
-
-data class PaintColor(val color: Color, val isChecked: Boolean)
