+++ app/src/main/java/woowacourse/paint/action/PaintActionHandler.kt
@@ -1,0 +1,9 @@
+package woowacourse.paint.action
+
+import androidx.annotation.ColorRes
+
+interface PaintActionHandler {
+    fun changeColorRes(
+        @ColorRes colorRes: Int,
+    )
+}
