+++ app/src/main/java/woowacourse/paint/adapter/PaintColorViewHolder.kt
@@ -1,7 +1,9 @@
-package woowacourse.paint
+package woowacourse.paint.adapter
 
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.paint.action.PaintActionHandler
 import woowacourse.paint.databinding.ItemPaintColorBinding
+import woowacourse.paint.util.Color
 
 class PaintColorViewHolder(private val binding: ItemPaintColorBinding) :
     RecyclerView.ViewHolder(binding.root) {
