+++ app/src/main/java/woowacourse/paint/model/Line.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.model
 
 import android.graphics.Paint
 import android.graphics.Path
