+++ app/src/main/java/woowacourse/paint/model/PaintColor.kt
@@ -1,0 +1,5 @@
+package woowacourse.paint.model
+
+import woowacourse.paint.util.Color
+
+data class PaintColor(val color: Color, val isChecked: Boolean)
