+++ app/src/main/java/woowacourse/paint/util/PaintBindingAdapter.kt
@@ -1,10 +1,11 @@
-package woowacourse.paint
+package woowacourse.paint.util
 
 import android.content.res.ColorStateList
 import android.widget.ImageView
 import androidx.annotation.ColorRes
 import androidx.core.content.ContextCompat
 import androidx.databinding.BindingAdapter
+import woowacourse.paint.R
 
 @BindingAdapter("isChecked")
 fun ImageView.bindPaintColorIsChecked(isChecked: Boolean) {
@@ -16,7 +17,9 @@ fun ImageView.bindPaintColorIsChecked(isChecked: Boolean) {
 }
 
 @BindingAdapter("colorBackgroundTint")
-fun ImageView.bindColorBackgroundTint(@ColorRes colorRes: Int) {
+fun ImageView.bindColorBackgroundTint(
+    @ColorRes colorRes: Int,
+) {
     val color = ContextCompat.getColor(context, colorRes)
     backgroundTintList = ColorStateList.valueOf(color)
 }
