+++ app/src/main/java/woowacourse/paint/viewmodel/PaintViewModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.viewmodel
 
 import androidx.annotation.ColorRes
 import androidx.lifecycle.LiveData
@@ -6,6 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import woowacourse.paint.PaintBoard.Companion.DEFAULT_PAINT_COLOR_RES
 import woowacourse.paint.PaintBoard.Companion.DEFAULT_STROKE_WIDTH
+import woowacourse.paint.action.PaintActionHandler
 
 class PaintViewModel : ViewModel(), PaintActionHandler {
     private val _colorRes: MutableLiveData<Int> = MutableLiveData(DEFAULT_PAINT_COLOR_RES)
@@ -14,7 +15,9 @@ class PaintViewModel : ViewModel(), PaintActionHandler {
     private val _strokeWidth: MutableLiveData<Float> = MutableLiveData(DEFAULT_STROKE_WIDTH)
     val strokeWidth: LiveData<Float> get() = _strokeWidth
 
-    override fun changeColorRes(@ColorRes colorRes: Int) {
+    override fun changeColorRes(
+        @ColorRes colorRes: Int,
+    ) {
         _colorRes.value = colorRes
     }
